C++ Classes
- Added C++ Class for the powerup spawner, powerup
- reAdded HUD for player
- Fixed issue with Hud not displaying correct powerup

+++ Source/project_clean/MyDirt.h
@@ -33,7 +33,7 @@ public:
 	/* Variables */
 	// Points
 	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = "Variables")
-	float Points;
+	int32 Points;
 
 	/* Functions */
 	UFUNCTION(BlueprintCallable, Category = "Dirt")
